blkid: enable evaluate by scan
Device tokens (UUID etc.) are resolved using either udev by-[token]
symlinks or by scanning the disks.

There's a race condition between when udev symlinks are created and
when systemd runs systemd-remount-fs. The race is triggered if
/etc/fstab uses UUIDs to identify partitions.

Enable evaluate by scan as a fallback if udev symlinks are not found.

Fixes: TIVI-1121

Signed-off-by: Mikko Ylinen <mikko.ylinen@intel.com>
Signed-off-by: William Douglas <william.douglas@intel.com>
Change-Id: Ic8c57cb542de59f1413f0ffebce2408ac7420623